requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.1.4
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
gunicorn==21.2.0
torch==2.1.1
transformers==4.36.2
